+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- feat: introduce shared backend-selection constructor NewBackend mapping resolved execution policy to DefaultBackend or PiAdapter
 ### Changed
 ### Fixed
 ### Removed